Refactor all names to use typescript instead of js
