--- codegen-server-test/js/build.gradle.kts
@@ -1,60 +1,0 @@
-/*
- * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-
-description = "Generates Rust/Js code from Smithy models and runs the protocol tests"
-extra["displayName"] = "Smithy :: Rust :: Codegen :: Server :: Js :: Test"
-extra["moduleName"] = "software.amazon.smithy.rust.kotlin.codegen.server.js.test"
-
-tasks["jar"].enabled = false
-
-plugins {
-    id("software.amazon.smithy")
-}
-
-val smithyVersion: String by project
-val defaultRustDocFlags: String by project
-val properties = PropertyRetriever(rootProject, project)
-
-val pluginName = "rust-server-codegen-js"
-val workingDirUnderBuildDir = "smithyprojections/codegen-server-test-js/"
-
-configure<software.amazon.smithy.gradle.SmithyExtension> {
-    outputDirectory = file("$buildDir/$workingDirUnderBuildDir")
-}
-
-buildscript {
-    val smithyVersion: String by project
-    dependencies {
-        classpath("software.amazon.smithy:smithy-cli:$smithyVersion")
-    }
-}
-
-dependencies {
-    implementation(project(":codegen-server:js"))
-    implementation("software.amazon.smithy:smithy-aws-protocol-tests:$smithyVersion")
-    implementation("software.amazon.smithy:smithy-protocol-test-traits:$smithyVersion")
-    implementation("software.amazon.smithy:smithy-aws-traits:$smithyVersion")
-}
-
-val allCodegenTests = "../../codegen-core/common-test-models".let { commonModels ->
-    listOf(
-        CodegenTest("com.amazonaws.simple#SimpleService", "simple", imports = listOf("$commonModels/simple.smithy")),
-        CodegenTest("com.aws.example.js#PokemonService", "pokemon-service-server-sdk"),
-    )
-}
-
-project.registerGenerateSmithyBuildTask(rootProject, pluginName, allCodegenTests)
-project.registerGenerateCargoWorkspaceTask(rootProject, pluginName, allCodegenTests, workingDirUnderBuildDir)
-project.registerGenerateCargoConfigTomlTask(buildDir.resolve(workingDirUnderBuildDir))
-
-tasks["smithyBuildJar"].dependsOn("generateSmithyBuild")
-tasks["assemble"].finalizedBy("generateCargoWorkspace")
-
-project.registerModifyMtimeTask()
-project.registerCargoCommandsTasks(buildDir.resolve(workingDirUnderBuildDir), defaultRustDocFlags)
-
-tasks["test"].finalizedBy(cargoCommands(properties).map { it.toString })
-
-tasks["clean"].doFirst { delete("smithy-build.json") }
