+++ codegen-server-test/typescript/model/pokemon-common.smithy
