+++ codegen-server-test/typescript/model/pokemon.smithy
