+++ codegen-server/typescript/src/main/resources/META-INF/services/software.amazon.smithy.build.SmithyBuildPlugin
@@ -1,0 +1,5 @@
+#
+# Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+software.amazon.smithy.rust.codegen.server.typescript.smithy.TsCodegenServerPlugin
