+++ rust-runtime/Cargo.toml
@@ -16,5 +16,5 @@ members = [
     "aws-smithy-xml",
     "aws-smithy-http-server",
     "aws-smithy-http-server-python",
-    "aws-smithy-http-server-js",
+    "aws-smithy-http-server-typescript",
 ]
