--- rust-runtime/aws-smithy-http-server-js/Cargo.toml
@@ -1,16 +1,0 @@
-[package]
-name = "aws-smithy-http-server-js"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-aws-smithy-http-server = { path = "../aws-smithy-http-server", features = ["aws-lambda", "shared-socket"] }
-http = "0.2"
-hyper = { version = "0.14", features = ["server", "http2", "tcp"]}
-napi = "2.10"
-napi-derive = "2.10"
-socket2 = "0.4"
-tokio = { version = "1.25" }
-tower = { version  = "0.4"}
