+++ settings.gradle.kts
@@ -10,10 +10,10 @@ include(":codegen-client")
 include(":codegen-client-test")
 include(":codegen-server")
 include(":codegen-server:python")
-include(":codegen-server:js")
+include(":codegen-server:typescript")
 include(":codegen-server-test")
 include(":codegen-server-test:python")
-include(":codegen-server-test:js")
+include(":codegen-server-test:typescript")
 include(":rust-runtime")
 include(":aws:sdk-codegen")
 include(":aws:sdk-adhoc-test")
